<feat> remove unused checkpoint

+++ tuner/cuhk_prototype_tuner_v2/cuhk_prototype_tuner_v2/utils.py
@@ -4,7 +4,33 @@ import copy
 import shutil
 import glob
 import hashlib
-heap=[]
+import json
+from collections import defaultdict
+
+def get_hash_code_from_config(config: dict):
+	obj = copy.deepcopy(config)
+	if 'TRIAL_BUDGET' in obj.keys():
+		obj.pop('TRIAL_BUDGET')
+	if "NUM_TRIAL_NEXT_ROUND" in obj.keys():
+		obj.pop("NUM_TRIAL_NEXT_ROUND")
+	hash_code = hashlib.md5(str(obj).encode('utf-8')).hexdigest()[:16]
+	return hash_code
+
+def delete_last_line(path):
+	with open(path, "rb+") as f:
+		try:
+			f.seek(-1, os.SEEK_END)
+			while f.read() != b"\n":
+				f.seek(-1, 1)
+				f.truncate()
+				f.seek(-1, 1)
+			f.seek(-1, 1)
+			f.truncate()
+		except IOError as e:
+			if e.errno == 22:
+				print("[INFO] File pointer reaches the head of file.")
+			else:
+				raise
 
 def preprocess(tid: str, config: dict, path: str):
 	'''
@@ -28,9 +54,7 @@ def preprocess(tid: str, config: dict, path: str):
 	------------
 	Maintain the checkpoints in the path based on configuration.
 	'''
-	item = copy.deepcopy(config)
-	item.pop('TRIAL_BUDGET')
-	hash_code = hashlib.md5(str(item).encode('utf-8')).hexdigest()[:16]
+	hash_code = get_hash_code_from_config(config)
 	load_pattern = '*' + hash_code + '*'
 	load_paths = glob.glob(os.path.join(path, load_pattern))
 	trial_budget = int(config['TRIAL_BUDGET'])
@@ -46,11 +70,14 @@ def preprocess(tid: str, config: dict, path: str):
 	return is_load, load_path, save_path, remain_trial_budget
 
 
-def postprocess(model, result, config, path, heap_size):
+def postprocess(eid, tid, result, config, path):
 	'''
 	Parameters
 	----------
-	model: tensorflow.python.keras.engine.training.Model
+	eid: str
+		experiment id
+	tid: str
+		trial id
 	result: dict
 		Invalid Value:
 			0.2
@@ -61,48 +88,56 @@ def postprocess(model, result, config, path, heap_size):
 			{'default':0.2,'attr0':0.3,'attr1':10,'maximize':'attr0'}
 			{'default':0.2,'attr0':0.3,'attr1':23.45,'maximize':['attr0','attr1']}
 	config: dict
-		trial configuration
+		trial configuration, must contain keys 'NUM_TRIAL_NEXT_ROUND' and 'TRIAL_BUDGET'
 	path: str
 		a path storing checkpointing folders 
-	heap_size: int
-		maximum number of ckpt to keep for next round's loading
 	
 	Return
 	------
-	None
+	is_save: bool
 
 	Introduction
 	------------
-	Maintain a heap of checkpoints with some given size.
+	Maintain a heap of checkpoints with a given size.
 
-	Notes
-	-----
-	Not compatible with PyTorch for the time being.
 	'''
-	if not isinstance(heap_size, int) or heap_size <= 0:
-		raise ValueError(f"ValueError: heap_size should be a positive integer but gets a value {heap_size}.")
-	global heap
-	if len(heap) == heap_size:
-		if hasattr(result, 'maximize')\
-		 and (isinstance(result['maximize'], str) and result['maximize'] == 'default'\
-		 or isinstance(result['maximize'], list) and 'default' in result['maximize']):
-			compared_item = heapq.nsmallest(1,heap,key=lambda x:x[0]['default'])[0]
-			if compared_item['default'] >= result['default']:
-				return
+	heap_size = int(config['NUM_TRIAL_NEXT_ROUND']) # maximum number of ckpt to keep
+	i = str(config['TRIAL_BUDGET'])
+	assert isinstance(heap_size, int) and heap_size >= 0, f"Heap size should be a non-negative integer but get a value {heap_size}."
+	
+	record_path = os.path.join(path, eid+".json")
+	heap = defaultdict(list)
+	if os.path.exists(record_path):
+		with open(record_path, 'r') as f:
+			records = f.readline()
+			heap = defaultdict(list, json.loads(records))
+			max_key = int(max(heap.keys(), key=int))
+			if int(i) < max_key: # Assume that each round is run in series, need to use another metrics if it's not
+				heap = defaultdict(list)
+	
+	if len(heap[i]) >= heap_size and heap_size != 0:
+		if 'maximize' in result.keys():
+			if (isinstance(result['maximize'], str) and result['maximize'] == 'default' \
+			  or isinstance(result['maximize'], list) and 'default' in result['maximize']):
+				compared_item = heapq.nsmallest(1,heap[i],key=lambda x:x[0]['default'])[0]
+				if compared_item[0]['default'] >= result['default']:
+					return False
+			else:
+				raise ValueError(f"Unexpected value of {result}")
 		else:
-			compared_item = heapq.nlargest(1,heap,key=lambda x:x[0]['default'])[0]
-			if compared_item[0] <= result['default']:
-				return
+			compared_item = heapq.nlargest(1,heap[i],key=lambda x:x[0]['default'])[0]
+			if compared_item[0]['default'] <= result['default']:
+				return False
 		try:
 			if os.path.exists(compared_item[1]):
 				shutil.rmtree(compared_item[1])
 		except OSError as e:
 			print("Error: %s : %s" % (compared_item[1], e.strerror))
-		heap.remove(compared_item)
-	item = copy.deepcopy(config)
-	item.pop('TRIAL_BUDGET')
-	save_dir = "-".join(config['TRIAL_BUDGET'], hash(item))
-	save_path = os.path.join(path, save_dir)
-	model.save(save_path)
-	heapq.append((result, save_path))
+		heap[i].remove(compared_item)
 
+	hash_code = get_hash_code_from_config(config)
+	save_path = os.path.join(path, "-".join([tid, str(config['TRIAL_BUDGET']), hash_code]))
+	heap[i].append([result, save_path])
+	with open(record_path, 'w') as f:
+		f.write(str(dict(heap)).replace("\'", '\"'))
+	return True
